xmm_lums.py is now stable, and functionally complete (for now at least).
It now calculates the luminosities in the cluster frame of reference rather than the rest frame. This involves generating XMM images with redshifted energy ranges, as well as changing the limits in the XSPEC runs.

Outputs a top level results file, and takes many more entries in the config file to be told what the expected headers are in the sample file.

Sample radius column can be in different units, with astropy handling the conversions.

Checked with Paul, and Planck15 is the cosmology he uses so that is fine.

Have included an example config file in the Git repo.

Have created a README and given an overview of using the code, as well as expected inputs to the config file/sample file, and expected outputs.