Small updates
Cuts to the eregionanalyse files to determine which clusters get measurements are now less harsh, it is possible to have background subtracted countrate of 0+/-0 but a non-zero upper limit for some reason...

Have added a configuration option that gets the user to choose which instrument to generate for.

README.md and config file example have been updated to reflect changes.